pyproject.toml
[tool.poetry]
name = "unicloud"
version = "0.2.0"
description = "cloud storage utils package"
authors = ["Mostafa Farrag <moah.farrag@gmail.com>"]
readme = "README.md"


[tool.poetry.urls]

[tool.poetry.dependencies]
python = "^3.12"
loguru = "^0.7.2"
numpy = "^2.1.2"
pandas = "^2.2.0"
google-cloud-storage = "^2.1.0"
google-api-python-client = "^2.119.0"
boto3 = "^1.35.40"


[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-cov = "^5.0.0"
flake8 = "^7.0.0"
coverage = {extras = ["toml"], version = "^7.6.9"}
safety = "^3.2.8"
mypy = "^1.13.0"
typeguard = "^4.3.0"
xdoctest = {extras = ["colors"], version = "^1.2.0"}
pre-commit = "^4.0.1"
black = "^24.10.0"
flake8-bandit = "^4.1.1"
flake8-bugbear = "^24.8.19"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.3.0"
pep8-naming = "^0.14.1"
darglint = "^1.8.1"
reorder-python-imports = "^3.14.0"
pre-commit-hooks = "^5.0.0"
pre-commit-commit-msg-hooks = "^0.1.0"
moto = "^5.0.22"
flake8-pyproject = "^1.2.3"

[tool.poetry.group.docs.dependencies]
sphinx = "^8.1.3"
pydata-sphinx-theme = "^0.15.2"

#[tool.poetry.extras]
#docs = ["sphinx", "sphinx_rtd_theme", "sphinx-autodoc-typehints", "recommonmark", "myst-parser"]
#sphinx = "^7.2.6"
[tool.poetry.scripts]


[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["unicloud"]

[tool.coverage.report]
show_missing = true
fail_under = 10

[tool.isort]
profile="black"
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
line_length=88
ensure_newline_before_comments=true


[tool.mypy]
strict = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.flake8]
ignore = ["E501", "E203", "E266", "W503"]
per-file-ignores = [
    '__init__.py:F401',
]
max-line-length = 88
count = true
max-complexity = 18
#select = B,C,E,F,W,T4


markers = [
    "fast: marks tests as fast (deselect with '-m \"not fast\"')",
    "e2e: marks tests as end-to-end (deselect with '-m \"e2e\"')",
    "serial",
    "mock: marks tests as mock (deselect with '-m \"mock\"')",
]

[build-system]
requires = ["setuptools", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
